Refactor server to use controllers, creating greater readability and separation of concerns
